requirements.txt
fastapi
uvicorn
python-multipart
sap-ai-sdk-gen[all]
openai
python-dotenv